fix #23447:Fatal "crash" attempting to delete all "Outer" members & then multipolygon - revert changes from r18967 - simply close relation editor dialog for the already deleted relation instead of generating an additional delete command
git-svn-id: https://josm.openstreetmap.de/svn/trunk@18968 0c6e7542-c601-0410-84e7-c038aed88b3b

+++ src/org/openstreetmap/josm/gui/dialogs/relation/actions/DeleteCurrentRelationAction.java
@@ -8,7 +8,10 @@ import java.beans.PropertyChangeEvent;
 
 import org.openstreetmap.josm.actions.mapmode.DeleteAction;
 import org.openstreetmap.josm.data.osm.Relation;
+import org.openstreetmap.josm.gui.MainApplication;
 import org.openstreetmap.josm.gui.dialogs.relation.GenericRelationEditor;
+import org.openstreetmap.josm.gui.dialogs.relation.RelationDialogManager;
+import org.openstreetmap.josm.gui.layer.OsmDataLayer;
 import org.openstreetmap.josm.tools.ImageProvider;
 
 /**
@@ -35,13 +38,20 @@ public class DeleteCurrentRelationAction extends AbstractRelationEditorAction {
         Relation toDelete = getEditor().getRelation();
         if (toDelete == null)
             return;
+        if (toDelete.isDeleted()) {
+            // see #23447
+            OsmDataLayer layer = MainApplication.getLayerManager().getEditLayer();
+            if (layer != null) {
+                RelationDialogManager.getRelationDialogManager().close(layer, toDelete);
+            }
+            return;
+        }
         DeleteAction.deleteRelation(getLayer(), toDelete);
     }
 
     @Override
-    public void updateEnabledState() {
-        Relation r = getEditor().getRelation();
-        setEnabled(r != null && !r.isDeleted());
+    protected void updateEnabledState() {
+        setEnabled(getEditor().getRelationSnapshot() != null);
     }
 
     @Override
